+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [7.0.3] - 2021-07-30
+### Added
+- Support for YAML examples
+- Infrastructure to easily add more supported formats
+
 ## [7.0.2] - 2021-04-03
 ### Fixed
 - Fixed License
