+++ src/Swashbuckle.AspNetCore.Filters/Examples/Formats/JsonExampleFormat.cs
@@ -1,0 +1,7 @@
+﻿namespace Swashbuckle.AspNetCore.Filters.Examples
+{
+    internal class JsonExampleFormat : ExampleFormat
+    {
+        public JsonExampleFormat() : base("application/json") { }
+    }
+}
